requirements.txt
certifi==2019.9.11
chardet==3.0.4
charset-normalizer==2.0.3
cloudinary==1.26.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==2.2.24
django-bootstrap3==15.0.0
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-registration==3.2
django-tinymce==3.3.0
django-url-or-relative-url-field==0.1.2
djangorestframework==3.11.2
fontawesome-free==5.15.3
gunicorn==19.9.0
idna==2.8
Pillow==8.2.0
psycopg2==2.8.3
python-dateutil==2.8.0
python-decouple==3.1
python-dotenv==0.10.3
pytz==2019.3
pyuploadcare==2.6.0
requests==2.26.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.26.6
whitenoise==4.1.4
testresources==2.0.1